Backend sell buy

+++ atelier2/Dynamique/SPCardGame/src/main/java/com/roby/oui/SPCardGame/model/CardFormDTO.java
@@ -18,7 +18,7 @@ public class CardFormDTO {
     private int energy;
     private int attack;
     private int defence;
-    private float prix;
+    private int prix;
 
 
     public CardFormDTO() {
